♻️ ASTをVirtual Nodeに変換する処理を分離
ASTをVirtual Nodeに変換する処理を、src/compiler/ast-to-vnode.mjs ファイルに移動し、外部ファイル化した。
